+++ libft/ft_lstnew_bonus.c
@@ -6,14 +6,14 @@
 /*   By: prambaud <prambaud@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/08 10:13:32 by prambaud          #+#    #+#             */
-/*   Updated: 2024/11/30 15:01:21 by prambaud         ###   ########.fr       */
+/*   Updated: 2024/12/02 09:52:51 by prambaud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-t_list	*ft_lstnew(int content)
+t_list	*ft_lstnew(long long content)
 {
 	t_list	*new;
